everything working 100%

+++ JavaProject/src/lab1/model/SortEdgesByWeight.java
@@ -5,7 +5,7 @@ import java.util.Comparator;
 public class SortEdgesByWeight implements Comparator<Edge> {
     @Override
     public int compare(Edge arg0, Edge arg1) {
-	return Integer.compare(arg0.getWeight().intValue(), 
-		arg1.getWeight().intValue());
+	return Integer.compare(arg0.getWeight(), 
+		arg1.getWeight());
     }
 }
